+++ framework/common/src/index.ts
@@ -1,4 +1,5 @@
 export * from './extractArgumentNames';
+export * from './getEnvVar';
 export * from './guards/isAsyncGenerator';
 export * from './guards/isNever';
 export * from './guards/isNodeJsWriteableStream';