+++ framework/core/src/services/index.ts
@@ -1,6 +1,6 @@
+export * from '../../../common/src/getEnvVar';
 export { createConfigInjector } from './configInjector';
 export type { ConfigInjector, ValidConfigInjector } from './configInjector';
-export * from './getEnvVar';
 export * from './interfaces/baseService';
 export * from './types/configInjector.types';
 export * from './types/entityManager.types';