+++ framework/express/src/types/express.types.ts
@@ -95,6 +95,7 @@ export interface MiddlewareResponse<
       | 'statusCode'
       | 'sendStatus'
       | 'getHeaders'
+      | 'getHeader'
       | 'setHeader'
       | 'send'
       | 'json'