+++ framework/hyper-express/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @forklaunch/hyper-express
 
+## 0.5.8
+
+### Patch Changes
+
+- Move getEnvVar into common package and allow for cors options during application instantiation
+- Updated dependencies
+  - @forklaunch/common@0.3.6
+  - @forklaunch/core@0.8.8
+  - @forklaunch/validator@0.6.6
+
 ## 0.5.7
 
 ### Patch Changes