+++ framework/validator/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @forklaunch/validator
 
+## 0.6.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @forklaunch/common@0.3.6
+
 ## 0.6.5
 
 ### Patch Changes